requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.7
pandas>=2.0.0
plotly>=5.15.0
openai>=1.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
altair>=5.0.0
streamlit-authenticator>=0.2.3
